# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OrderFlow Platform API - Multi-tenant ordering platform with onboarding workflow
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiConciergeChat
      tags: [ai]
      summary: Send a message to the AI sales concierge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/catalog-search:
    post:
      operationId: aiCatalogSearch
      tags: [ai]
      summary: Semantic keyword search over the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCatalogSearchBody"
      responses:
        "200":
          description: Matching catalog items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCatalogSearchResponse"
  /ai/upsell:
    post:
      operationId: aiUpsellSuggestions
      tags: [ai]
      summary: Get AI upsell suggestions based on current cart or order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiUpsellBody"
      responses:
        "200":
          description: Upsell suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiUpsellResponse"

  # ─────────────── NOTIFICATIONS ───────────────
...
components:
  schemas:
...
    CatalogItem:
      type: object
      properties:
        sellableProduct:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            options:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  catalogItemId:
                    type: integer
                  label:
                    type: string
                  price:
                    type: string
                  sku:
                    type: string
                    nullable: true
                required: [id, catalogItemId, label, price]
          required: [id, name, options]
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        stockQuantity:
          type: integer
        isAvailable:
          type: boolean
        isTaxable:
          type: boolean
        imageUrl:
          type: string
        mediaGallery:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [image, video]
              src:
                type: string
              alt:
                type: string
                nullable: true
            required: [type, src]
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        isFeatured:
          type: boolean
        isSaleFeatured:
          type: boolean
        alavontName:
          type: string
          nullable: true
        alavontDescription:
          type: string
          nullable: true
        alavontCategory:
          type: string
          nullable: true
        alavontImageUrl:
          type: string
          nullable: true
        alavontInStock:
          type: boolean
          nullable: true
        luciferCruzName:
          type: string
          nullable: true
        luciferCruzImageUrl:
          type: string
          nullable: true
        luciferCruzDescription:
          type: string
          nullable: true
        luciferCruzCategory:
          type: string
          nullable: true
        regularPrice:
          type: number
          nullable: true
        homiePrice:
          type: number
          nullable: true
        merchantProcessingMode:
          type: string
          nullable: true
        merchantProductSource:
          type: string
          nullable: true
        isWooManaged:
          type: boolean
        isLocalAlavont:
          type: boolean
        wooProductId:
          type: string
          nullable: true
        wooVariationId:
          type: string
          nullable: true
        receiptName:
          type: string
          nullable: true
        parLevel:
          type: number
          nullable: true
          minimum: 0
        moq:
          type: number
          minimum: 0
        preferredReorderQuantity:
          type: number
          minimum: 0
        lifecycleStatus:
          type: string
          enum: [customer_visible, unavailable_hidden, compliance_hold, archived]
        labName:
          type: string
          nullable: true
        customerSafeName:
          type: string
          nullable: true
        customerSafeDescription:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, name, category, price, isAvailable, createdAt, updatedAt]
...
    AiChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]
    CartAction:
      type: object
      properties:
        action:
          type: string
          enum: [add, remove, update_quantity]
        catalogItemId:
          type: integer
        quantity:
          type: integer
        itemName:
          type: string
      required: [action, catalogItemId]
    AiCartItem:
      type: object
      properties:
        catalogItemId:
          type: integer
        quantity:
          type: integer
        name:
          type: string
        price:
          type: number
      required: [catalogItemId, quantity]
    AiChatBody:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiChatMessage"
        cart:
          type: array
          items:
            $ref: "#/components/schemas/AiCartItem"
        context:
          type: object
          properties:
            currentCartItems:
              type: array
              items:
                type: integer
            currentOrderId:
              type: integer
      required: [messages]
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        suggestedItems:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
        cartActions:
          type: array
          items:
            $ref: "#/components/schemas/CartAction"
        conversationId:
          type: string
      required: [reply]
    AiCatalogSearchBody:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
      required: [query]
    AiCatalogSearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
        query:
          type: string
      required: [items, query]
    AiUpsellBody:
      type: object
      properties:
        cartItemIds:
          type: array
          items:
            type: integer
        orderId:
          type: integer
      required: [cartItemIds]
    AiUpsellResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
        reasoning:
          type: string
      required: [suggestions]

    # ─── Notifications ───
...